requirements.txt
spacy==3.7.4
textstat==0.7.4
nltk==3.8.1
python-dotenv==1.0.1
openai==1.35.9
